rbd: remove features_nautilus.go
